#356 Merge BePaid classes into one
